ashmem: Allow for restricting mapping permissions multiple times
The existing logic applies F_SEAL_SEAL when applying
F_SEAL_FUTURE_WRITE. This makes it so that no other seals can be applied
on a file.

This could be problematic in the future when executable mappings are
sealed through their own separate seal. For example, setting the buffer
to be mappable as RW would add the executable seal, and changing the
buffer to be mappable as RO would add the write seal. This would fail
if the first seal addition prevented any new seals from being added.

Therefore, only add seals when setting the protection mask for a buffer,
and do not prevent the addition of new seals.

Bug: 111903542
Change-Id: I1b765ba784991d5a3f4d1b866016cb17d3a33aaa
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>

Former-commit-id: 2a760959c48682245c1236219f2e0dbc1c0377a5