Add fs_config entry for first_stage.sh
First stage init will not run first_stage.sh if it's not executable,
even if androidboot.first_stage_init=1. If mkbootfs is used to package
the ramdisk, this results in a ramdisk with a first_stage.sh that isn't
executable. Add default entry for first_stage.sh with executable bits
set.

Bug: 386837572
Change-Id: Ic7f50141cca678003177252dc760d06c39f91e80
Signed-off-by: Elliot Berman <elliot.berman@oss.qualcomm.com>

Former-commit-id: f97da3c6a8e0e7cb0dc8a03640aa99bfcdf34ea7