Fix potential flaky debuggerd tests.
Some of the tests rely on the fact that a map created for the test
will be listed in the maps part of a tombstone as distinct. However,
if there was an unnamed map that starts right before the test map,
it appears in the tombstone map list as a single larger map and
causes tests to fail. To fix this, name the maps with prctl to
guarantee that these are distinct and will not result in a
combined map in the tombstone.

Bug: 395295765

Test: All unit tests pass.
Change-Id: Ib8a7c6dbfbc810f67fe5e5b9cb7802bbbec2cbba

Former-commit-id: 9c4b0eeea83d64788c91c3619f86f0836c89dfef