ashmem: Expose has_memfd_support() for ashmem-tests to use
has_memfd_support() can be used to determine if memfds support ashmem
ioctl commands. This will be useful for the ashmem test suite, as it
can use that information to determine if it should test the
implementation of the memfd-ashmem ioctls.

Therefore, expose has_memfd_support() but only for ashmem-tests to use.
Ashmem clients do not need/must not use this function.

Bug: 111903542
Change-Id: If1378f73ddbba9a94c45375357ebf56fb0170b51
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>

Former-commit-id: 78541b718d631348eb494e18ca55be219bd709dc