Remove obsolete configs from fs_config.cpp
The removed lines are for configuring directories or files in the
flattened APEXes. Since flattened APEX no longer exists, these lines are
obsolete, and thus removing.

Ownership and mode of files/dirs in APEXes are configured via the
canned_fs_config files specific to the APEXes.

Bug: 402962730
Test: build the tethering APEX, check the owner ahd mode of the
for-system directory in it.

Change-Id: I58e9b67f5cedd8233e02b4d2f4d9c277d6f64f8c

Former-commit-id: 9cdca9a61d484b7defea4101c0eb3e90aff779b2